Route-finding: enemies consider NPCs when finding a route to player. NB: NPCs should be placed at the centre of a tile by default.

+++ core/src/main/java/com/genericgames/samurai/model/Door.java
@@ -5,7 +5,7 @@ public class Door extends WorldObject implements Collidable {
     private String fileName;
     private int spawnNumber;
 
-    public Door(int positionX, int positionY){
+    public Door(float positionX, float positionY){
         super(positionX, positionY);
     }
