+++ core/src/main/java/com/genericgames/samurai/model/SpawnPoint.java
@@ -4,7 +4,7 @@ public class SpawnPoint extends WorldObject {
 
     int spawnNumber;
 
-    public SpawnPoint(int posX, int posY, int spawnNumber){
+    public SpawnPoint(float posX, float posY, int spawnNumber){
         super(posX, posY);
         this.spawnNumber = spawnNumber;
     }
